try to handle signal -> to fix

+++ src/signal_handler.c
@@ -6,7 +6,7 @@
 /*   By: vsporer <vsporer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/13 14:02:25 by vsporer           #+#    #+#             */
-/*   Updated: 2017/10/13 15:47:02 by vsporer          ###   ########.fr       */
+/*   Updated: 2017/10/15 15:28:13 by vsporer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,10 @@ void	signal_handler(int sig)
 	pid_t	last;
 
 	last = last_pid(0);
-	if (sig == SIGINT && last)
-		kill(last, SIGINT);
+	signal_value(sig);
+	if (last > 0)
+	{
+		if (sig == SIGINT && last)
+			kill(last, SIGINT);
+	}
 }
